[argparse] Prepare argparse for contributing to OTP
Refactor APIs so exceptions are thrown only when
there is an invalid command specification passed.

Preparing argparse for OTP: now exceptions are used
only when there is a developer error (badarg), and
parser returns {error, Reason} tuple when there is
a parser error.

Added basic run/3 function in the argparse itself,
because cli.erl is not generic enough for majority
of use-cases.